refactor: (#197) Converter 오타 수정
