+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/file/convertor/ExcelFileConverter.kt
@@ -1,18 +1,18 @@
-package team.comit.simtong.domain.file.transfer
+package team.comit.simtong.domain.file.convertor
 
 import org.springframework.web.multipart.MultipartFile
-import team.comit.simtong.domain.file.FileExtensionUtils.XLS
-import team.comit.simtong.domain.file.FileExtensionUtils.XLSX
+import team.comit.simtong.domain.file.convertor.FileExtensions.XLS
+import team.comit.simtong.domain.file.convertor.FileExtensions.XLSX
 
 /**
  *
- * Excel File의 변환을 담당하는 ExcelFileConvertor
+ * Excel File의 변환을 담당하는 ExcelFileConverter
  *
  * @author Chokyunghyeon
  * @date 2022/12/09
  * @version 1.0.0
  **/
-object ExcelFileConvertor : FileConvertor {
+object ExcelFileConverter : FileConverter {
 
     override fun isCorrectExtension(multipartFile: MultipartFile): Boolean {
         return when (multipartFile.extension) {
